Move hide_qt_warning() to qtlog
Just to keep related things together.
