+++ tests/unit/utils/test_log.py
@@ -340,35 +340,6 @@ class TestInitLog:
         assert log.console_filter.names == {'misc'}
 
 
-class TestHideQtWarning:
-
-    """Tests for hide_qt_warning/QtWarningFilter."""
-
-    @pytest.fixture
-    def qt_logger(self):
-        return logging.getLogger('qt-tests')
-
-    def test_unfiltered(self, qt_logger, caplog):
-        with log.hide_qt_warning("World", 'qt-tests'):
-            with caplog.at_level(logging.WARNING, 'qt-tests'):
-                qt_logger.warning("Hello World")
-        assert len(caplog.records) == 1
-        record = caplog.records[0]
-        assert record.levelname == 'WARNING'
-        assert record.message == "Hello World"
-
-    @pytest.mark.parametrize('line', [
-        "Hello",  # exact match
-        "Hello World",  # match at start of line
-        "  Hello World  ",  # match with spaces
-    ])
-    def test_filtered(self, qt_logger, caplog, line):
-        with log.hide_qt_warning("Hello", 'qt-tests'):
-            with caplog.at_level(logging.WARNING, 'qt-tests'):
-                qt_logger.warning(line)
-        assert not caplog.records
-
-
 @pytest.mark.parametrize('suffix, expected', [
     ('', 'STUB: test_stub'),
     ('foo', 'STUB: test_stub (foo)'),
